Split root command black-box tests
