+++ .teststyle-baseline.json
@@ -1149,11 +1149,6 @@
       "reason": "same-package test file is not named *_internal_test.go"
     },
     {
-      "path": "cmd/root/root_test.go",
-      "package": "root",
-      "reason": "same-package test file is not named *_internal_test.go"
-    },
-    {
       "path": "core/goschema/embedded_bug_reproduction_test.go",
       "package": "goschema",
       "reason": "same-package test file is not named *_internal_test.go"
